# backend-dotnet/MapService/OpenAPISpecification/api.yml
openapi: 3.0.3
info:
  title: hajk-backend
  description: Hajk backend
  version: 1.1.0
servers: []
...
paths:
...
  /config/layers:
    get:
      tags:
        - Client-accessible
      description: List available layers. If AD authentication is active, filter by user's permission.
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: All layers were fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigLayers"
...
  /config/list:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: List available map configurations
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Available maps were fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigList"
...
  /config/listimage:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: List available images in the upload folder
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Available images were fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListImage"
...
  /config/listvideo:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: List available videos in the upload folder
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Available videos were fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListVideo"
...
  /config/listaudio:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: List available audio files in the upload folder
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Available audio files were fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAudio"
...
  /config/{map}:
    get:
      deprecated: false
      tags:
        - Client-accessible
      description: Fetch contents of a map configuration
      parameters:
        - name: map
          in: path
          description: Name of the map configuration to fetch
          required: true
          schema:
            type: string
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Success
          content: {}
...
  /config/create/{name}:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: Create a new map configuration
      parameters:
        - name: name
          in: path
          description: File name of the new map configuration
          required: true
          schema:
            type: string
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Success
          content: {}
...
  /config/delete/{name}:
    get:
      deprecated: true
      tags:
        - Admin - Maps and layers
      description: Delete an existing map configuration
      parameters:
        - name: name
          in: path
          description: Name of the map configuration to be deleted
          required: true
          schema:
            type: string
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Success
          content: {}
...
  /config/userspecificmaps:
    get:
      tags:
        - Client-accessible
      description: Get a list of maps that will be available in the MapSwitcher dropdown
      parameters:
        - in: header
          name: X-Control-Header
          description: User name that will be supplied to AD
          schema:
            type: string
          required: false
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSpecificMaps"
...
components:
  schemas:
...
    ConfigList:
      type: object
      properties:
        "0":
          type: string
          example: map_1
        "1":
          type: string
          example: nytt
        "2":
          type: string
          example: oversiktsplan
        "3":
          type: string
          example: oversiktsplan_OLD
    ListImage:
      type: array
      items:
        type: string
        example: geografiska_inriktningar_1.PNG
    ListAudio:
      type: array
      items:
        type: string
        example: mixkit-arcade-retro-game-over-213.wav
    ListVideo:
      type: array
      items:
        type: string
        example: file_example_OGG_480_1_7mg.ogg
...
    UserSpecificMaps:
      type: object
      properties:
        "0":
          type: object
          properties:
            mapConfigurationName:
              type: string
              example: map_1
            mapConfigurationTitle:
              type: string
              example: Kartan Map 1
        "1":
          type: object
          properties:
            mapConfigurationName:
              type: string
              example: nytt
            mapConfigurationTitle:
              type: string
              example: Hajk 3 template
...
    ConfigLayers:
      type: object
      properties:
        wmtslayers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 1
              caption:
                type: string
                example: Topografiska webbkartan öppna data1
              internalLayerName:
                type: string
                format: nullable
              url:
                type: string
                example: https://api.lantmateriet.se/open/topowebb-ccby/v1/wmts/token/b84e03e0-c917-313e-9f7d-fb5303034148/
              owner:
                type: string
                format: nullable
              date:
                type: string
                example: 1664201882947
              content:
                type: string
                example: Topowebbkartan i WMTS-format
              legend:
                type: string
                example: null
              legendIcon:
                type: string
                example: null
              projection:
                type: string
                example: EPSG:3006
              visibleAtStart:
                type: boolean
                description: Show at start
              layer:
                type: string
                example: topowebb_nedtonad
              style:
                type: string
                example: default
              matrixSet:
                type: string
                example: 3006
              origin:
                type: array
                items:
                  type: string
                  example: -1200000
              resolutions:
                type: array
                items:
                  type: string
                  example: 2048
              matrixIds:
                type: array
                items:
                  type: string
                  example: 1
              attribution:
                type: string
                example: Lantmäteriet
              infoVisible:
                type: boolean
              infoTitle:
                type: string
                example: Topowebbkartan i WMTS-format
              infoText:
                type: string
                example: Bakgrundskarta från Lantmäteriet i WMTS-format.
              infoUrl:
                type: string
                example: https://www.lantmateriet.se/globalassets/kartor-och-geografisk-information/geodatatjanster/tb_twkvisningv1.0.4.pdf
              infoUrlText:
                type: string
                example: Information om tjänsten
              infoOwner:
                type: string
                example: Lantmäteriet
              timeSliderVisible:
                type: boolean
              timeSliderStart:
                type: string
                example: null
              timeSliderEnd:
                type: string
                example: null
              infoClickSortType:
                type: string
                format: nullable
              infoClickSortDesc:
                type: boolean
              infoClickSortProperty:
                type: string
                format: nullable
              zIndex:
                type: string
                format: nullable
              minZoom:
                type: integer
                format: int32
                example: -1
              maxZoom:
                type: integer
                format: int32
                example: -1
        wmslayers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 2
              caption:
                type: string
                example: Nedtonad stadskarta Göteborg
              internalLayerName:
                type: string
                format: nullable
              url:
                type: string
                example: https://opengeodata.goteborg.se/services/stadskarta/wms/v4
              customGetMapUrl:
                type: string
                format: nullable
              owner:
                type: string
                example: null
              date:
                type: string
                example: 1580375228781
              content:
                type: string
                example: null
              legend:
                type: string
                example: null
              legendIcon:
                type: string
                format: nullable
              projection:
                type: string
                example: EPSG:3006
              layers:
                type: array
                items:
                  type: string
                  example: stadskarta_nedtonad
              layersInfo:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: stadskarta_nedtonad
                    caption:
                      type: string
                      example: null
                    internalLayerName:
                      type: string
                      format: nullable
                    legend:
                      type: string
                      example: null
                    legendIcon:
                      type: string
                      format: nullable
                    infobox:
                      type: string
                      example: null
                    style:
                      type: string
                      example: null
                    queryable:
                      type: boolean
                    searchDisplayName:
                      type: string
                      format: nullable
                    searchGeometryField:
                      type: string
                      format: nullable
                    searchOutputFormat:
                      type: string
                      format: nullable
                    searchPropertyName:
                      type: string
                      format: nullable
                    searchUrl:
                      type: string
                      format: nullable
              searchFields:
                type: string
                format: nullable
              displayFields:
                type: string
                format: nullable
              visibleAtStart:
                type: boolean
                description: Show at start
              hideExpandArrow:
                type: boolean
              tiled:
                type: boolean
              opacity:
                type: integer
                format: int32
                example: 1
              minZoom:
                type: integer
                format: int32
                example: 0
              maxZoom:
                type: integer
                format: int32
                example: 0
              minMaxZoomAlertOnToggleOnly:
                type: boolean
              singleTile:
                type: boolean
              imageFormat:
                type: string
                example: image/png
              serverType:
                type: string
                example: geoserver
              attribution:
                type: string
                example: null
              searchUrl:
                type: string
                example: null
              customRatio:
                type: integer
                format: int32
                example: 0
              hidpi:
                type: boolean
              searchPropertyName:
                type: string
                example: null
              searchDisplayName:
                type: string
                example: null
              searchOutputFormat:
                type: string
                example: null
              searchGeometryField:
                type: string
                example: null
              infoVisible:
                type: boolean
              infoTitle:
                type: string
                example: null
              infoText:
                type: string
                example: Sammansatt bakgrundskarta i gråton.
              infoUrl:
                type: string
                example: null
              infoUrlText:
                type: string
                example: null
              infoOwner:
                type: string
                example: Stadsbyggnadskontoret Göteborg
              infoClickSortType:
                type: string
                format: nullable
              infoClickSortDesc:
                type: boolean
              infoClickSortProperty:
                type: string
                format: nullable
              version:
                type: string
                example: 1.3.0
              infoFormat:
                type: string
                example: application/json
              timeSliderVisible:
                type: boolean
              timeSliderStart:
                type: string
                format: nullable
              timeSliderEnd:
                type: string
                format: nullable
              zIndex:
                type: string
                format: nullable
        wfslayers:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: 4
              caption:
                type: string
                example: Adresser Göteborg
              internalLayerName:
                type: string
                example: Adresser Göteborg
              url:
                type: string
                example: https://opengeodata.goteborg.se/services/adresser/wfs/v1
              layers:
                type: array
                items:
                  type: string
                  example: adresser.wfs.v1:adresser
              searchFields:
                type: array
                items:
                  type: string
                  example: adress
              infobox:
                type: string
                example: adress
              aliasDict:
                type: string
                example: null
              displayFields:
                type: array
                items:
                  type: string
                  example: adress
              geometryField:
                type: string
                example: geom
              outputFormat:
                type: string
                example: GML2
              infoVisible:
                type: boolean
              infoTitle:
                type: string
                format: nullable
              infoText:
                type: string
                format: nullable
              infoUrl:
                type: string
                format: nullable
              infoUrlText:
                type: string
                format: nullable
              infoOwner:
                type: string
                format: nullable
              zIndex:
                type: string
                format: nullable
        vectorlayers:
          type: array
          items:
            type: object
        wfstlayers:
          type: array
          items:
            type: object
        arcgislayers:
          type: array
          items:
            type: object
...